refactor: api 타입 분리
