+++ src/apis/movie/api.ts
@@ -1,78 +1,11 @@
 import { getSearchParamsFromObject } from "../../utils/getSearchParamsFromObject";
 import { tmdbFetcher, TmdbPagination } from "../../utils/tmdbFetcher";
-
-export interface Movie {
-  adult: boolean;
-  backdrop_path: string;
-  genre_ids: number[];
-  id: number;
-  original_language: string;
-  original_title: string;
-  overview: string;
-  popularity: number;
-  poster_path: string;
-  release_date: string;
-  title: string;
-  video: boolean;
-  vote_average: number;
-  vote_count: number;
-}
-
-export interface MovieDetail extends Omit<Movie, "genre_ids"> {
-  belongs_to_collection: BelongsToCollection;
-  budget: number;
-  genres: Genre[];
-  homepage: string;
-  imdb_id: string;
-  origin_country: string[];
-  production_companies: ProductionCompany[];
-  production_countries: ProductionCountry[];
-  revenue: number;
-  runtime: number;
-  spoken_languages: SpokenLanguage[];
-  status: string;
-  tagline: string;
-}
-
-export interface BelongsToCollection {
-  id: number;
-  name: string;
-  poster_path: string;
-  backdrop_path: string;
-}
-
-export interface Genre {
-  id: number;
-  name: string;
-}
-
-export interface ProductionCompany {
-  id: number;
-  logo_path: string;
-  name: string;
-  origin_country: string;
-}
-
-export interface ProductionCountry {
-  iso_3166_1: string;
-  name: string;
-}
-
-export interface SpokenLanguage {
-  english_name: string;
-  iso_639_1: string;
-  name: string;
-}
-
-export interface PopularMoviesParameter {
-  language: string;
-  page: number;
-}
-
-export interface MovieDetailParameter {
-  language: string;
-  movieId: number;
-}
+import {
+  Movie,
+  MovieDetail,
+  MovieDetailParameter,
+  PopularMoviesParameter,
+} from "./type.ts";
 
 export const getPopularMovies = async (
   params: Partial<PopularMoviesParameter> = {},
