+++ src/apis/search/type.ts
@@ -1,0 +1,5 @@
+export interface SearchedMoviesParameter {
+  query: string;
+  language: string;
+  page: number;
+}
