+++ src/dom/components/MainThumbnailList.ts
@@ -1,4 +1,4 @@
-import { Movie } from "../../apis/movie/api";
+import { Movie } from "../../apis/movie/type";
 import {
   renderMovieItems,
   renderMovieItemsLoading,
