+++ src/dom/components/SearchThumbnailList.ts
@@ -1,4 +1,4 @@
-import { Movie } from "../../apis/movie/api";
+import { Movie } from "../../apis/movie/type";
 import { renderMovieItems, renderMovieItemsLoading } from "../shared/MovieItem.ts";
 
 const SEARCH_THUMBNAIL_LIST_ID = "search-thumbnail-list";
