+++ src/pages/home.ts
@@ -1,4 +1,5 @@
-import { getPopularMovies, Movie } from "../apis/movie/api.ts";
+import { getPopularMovies } from "../apis/movie/api.ts";
+import { Movie } from "../apis/movie/type.ts";
 import TMDBError from "../TMDBError.ts";
 import { renderBanner } from "../dom/components/Banner.ts";
 import { renderMain, renderMainEmpty, renderMainError, renderMainLoading } from "../dom/compositions/Main.ts";
